pyproject.toml
[project]
name = "reposnap"
version = "0.3.2"
description = "Generate a Markdown file with all contents of your project"
authors = [
    { name = "agoloborodko" }
]
dependencies = [
    "pathlib>=1.0.1",
    "pathspec>=0.12.1",
    "gitpython>=3.1.43",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.3.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/reposnap"]

[project.scripts]
reposnap = "reposnap:main"
